Gun_1/requirements.txt
python-dotenv
requests
pandas
beautifulsoup4
selenium
webdriver-manager
firecrawl-py
